+++ src/Flutter.Material/ThemeData.cs
@@ -144,6 +144,7 @@ public sealed record ThemeData
     private OutlinedButtonThemeData? _outlinedButtonTheme;
     private FilledButtonThemeData? _filledButtonTheme;
     private IconButtonThemeData? _iconButtonTheme;
+    private DrawerThemeData? _drawerTheme;
     private FloatingActionButtonThemeData? _floatingActionButtonTheme;
     private BottomNavigationBarThemeData? _bottomNavigationBarTheme;
     private CheckboxThemeData? _checkboxTheme;
@@ -185,6 +186,7 @@ public sealed record ThemeData
         OutlinedButtonThemeData? outlinedButtonTheme = null,
         FilledButtonThemeData? filledButtonTheme = null,
         IconButtonThemeData? iconButtonTheme = null,
+        DrawerThemeData? drawerTheme = null,
         FloatingActionButtonThemeData? floatingActionButtonTheme = null,
         BottomNavigationBarThemeData? bottomNavigationBarTheme = null,
         CheckboxThemeData? checkboxTheme = null,
@@ -225,6 +227,7 @@ public sealed record ThemeData
         _outlinedButtonTheme = outlinedButtonTheme;
         _filledButtonTheme = filledButtonTheme;
         _iconButtonTheme = iconButtonTheme;
+        _drawerTheme = drawerTheme;
         _floatingActionButtonTheme = floatingActionButtonTheme;
         _bottomNavigationBarTheme = bottomNavigationBarTheme;
         _checkboxTheme = checkboxTheme;
@@ -322,6 +325,12 @@ public sealed record ThemeData
     {
         get => _iconButtonTheme ?? new IconButtonThemeData(style: IconButtonStyle);
         init => _iconButtonTheme = value;
+    }
+
+    public DrawerThemeData DrawerTheme
+    {
+        get => _drawerTheme ?? new DrawerThemeData();
+        init => _drawerTheme = value;
     }
 
     public FloatingActionButtonThemeData FloatingActionButtonTheme